+++ kwave/kWaveSimulation_helper/display_simulation_params.py
@@ -1,7 +1,7 @@
 import numpy as np
 
 from kwave import kWaveGrid, kWaveMedium
-from kwave.utils.conversion import scale_SI
+from kwave.utils.data import scale_SI
 
 
 def display_simulation_params(kgrid: kWaveGrid, medium: kWaveMedium, elastic_code: bool):
