+++ kwave/ksource.py
@@ -4,7 +4,7 @@ from warnings import warn
 import numpy as np
 
 from kwave.kgrid import kWaveGrid
-from kwave.utils.checks import num_dim2
+from kwave.utils.matrix import num_dim2
 
 
 @dataclass
