+++ kwave/utils/interp.py
@@ -5,13 +5,10 @@ from numpy.fft import fft, fftshift
 from scipy.interpolate import interpn
 from scipy.signal import resample
 
-from .conversion import scale_time, grid2cart
+from .conversion import grid2cart
+from .data import scale_time
+from .matrix import sort_rows
 from .tictoc import TicToc
-
-
-def sort_rows(arr: np.ndarray, index: int):
-    assert arr.ndim == 2, "'sort_rows' currently supports only 2-dimensional matrices"
-    return arr[arr[:, index].argsort()]
 
 
 def interpolate3d(grid_points: List[np.ndarray], grid_values: np.ndarray, interp_locs: List[np.ndarray]) -> np.ndarray:
@@ -146,155 +143,6 @@ def interpolate2d_with_queries(
         # set values outside the interpolation range to original values
         result[np.isnan(result)] = grid_values[np.isnan(result)]
     return result
-
-
-def cart2grid(kgrid, cart_data, axisymmetric=False):
-    """
-    Interpolate a set of Cartesian points onto a binary grid.
-
-    Args:
-        kgrid:
-        cart_data:
-        axisymmetric:
-
-    Returns:
-        cart2grid interpolates the set of Cartesian points defined by
-        cart_data onto a binary matrix defined by the kWaveGrid object
-        kgrid using nearest neighbour interpolation. An error is returned if
-        the Cartesian points are outside the computational domain defined by
-        kgrid.
-    """
-    # check for axisymmetric input
-    if axisymmetric and kgrid.dim != 2:
-        raise AssertionError('Axisymmetric flag only supported in 2D.')
-
-    # detect whether the inputs are for one, two, or three dimensions
-    if kgrid.dim == 1:
-        # one-dimensional
-        data_x = cart_data[0, :]
-
-        # scale position values to grid centered pixel coordinates using
-        # nearest neighbour interpolation
-        data_x = np.round(data_x / kgrid.dx).astype(int)
-
-        # shift pixel coordinates to coincide with matrix indexing
-        data_x = data_x + np.floor(kgrid.Nx // 2).astype(int)
-
-        # check if the points all lie within the grid
-        if data_x.max() > kgrid.Nx or data_x.min() < 1:
-            raise AssertionError('Cartesian points must lie within the grid defined by kgrid.')
-
-        # create empty grid
-        grid_data = np.zeros((kgrid.Nx, 1))
-
-        # create index variable
-        point_index = np.arange(1, data_x.size + 1)
-
-        # map values
-        for data_index in range(data_x.size):
-            grid_data[data_x[data_index]] = point_index[data_index]
-
-        # extract reordering index
-        reorder_index = np.reshape(grid_data[grid_data != 0], (-1, 1))
-
-    elif kgrid.dim == 2:
-        # two-dimensional
-        data_x = cart_data[0, :]
-        data_y = cart_data[1, :]
-
-        # scale position values to grid centered pixel coordinates using
-        # nearest neighbour interpolation
-        data_x = np.round(data_x / kgrid.dx).astype(int)
-        data_y = np.round(data_y / kgrid.dy).astype(int)
-
-        # shift pixel coordinates to coincide with matrix indexing (leave
-        # y-direction = radial-direction if axisymmetric)
-        data_x = data_x + np.floor(kgrid.Nx // 2).astype(int)
-        if not axisymmetric:
-            data_y = data_y + np.floor(kgrid.Ny // 2).astype(int)
-        else:
-            data_y = data_y
-
-        # check if the points all lie within the grid
-        if data_x.max() >= kgrid.Nx or data_y.max() >= kgrid.Ny or data_x.min() < 0 or data_y.min() < 0:
-            raise AssertionError('Cartesian points must lie within the grid defined by kgrid.')
-
-        # create empty grid
-        grid_data = -1 * np.ones((kgrid.Nx, kgrid.Ny))
-
-        # map values
-        for data_index in range(data_x.size):
-            grid_data[data_x[data_index], data_y[data_index]] = int(data_index)
-
-        # extract reordering index
-        reorder_index = grid_data.flatten(order='F')[
-            grid_data.flatten(order='F') != -1
-            ]
-        reorder_index = reorder_index[:, None] + 1  # [N] => [N, 1]
-
-    elif kgrid.dim == 3:
-
-        # three dimensional
-        data_x = cart_data[0, :]
-        data_y = cart_data[1, :]
-        data_z = cart_data[2, :]
-
-        # scale position values to grid centered pixel coordinates using
-        # nearest neighbour interpolation
-        data_x = np.round(data_x / kgrid.dx).astype(int)
-        data_y = np.round(data_y / kgrid.dy).astype(int)
-        data_z = np.round(data_z / kgrid.dz).astype(int)
-
-        # shift pixel coordinates to coincide with matrix indexing
-        data_x = data_x + np.floor(kgrid.Nx // 2).astype(int)
-        data_y = data_y + np.floor(kgrid.Ny // 2).astype(int)
-        data_z = data_z + np.floor(kgrid.Nz // 2).astype(int)
-
-        # check if the points all lie within the grid
-        assert 0 <= data_x.min() and 0 <= data_y.min() and 0 <= data_z.min() and \
-               data_x.max() < kgrid.Nx and data_y.max() < kgrid.Ny and data_z.max() < kgrid.Nz, \
-            "Cartesian points must lie within the grid defined by kgrid."
-
-        # create empty grid
-        grid_data = -1 * np.ones((kgrid.Nx, kgrid.Ny, kgrid.Nz), dtype=int)
-
-        # create index variable
-        point_index = np.arange(1, data_x.size + 1)
-
-        # map values
-        for data_index in range(data_x.size):
-            grid_data[data_x[data_index], data_y[data_index], data_z[data_index]] = point_index[data_index]
-
-        # extract reordering index
-        reorder_index = grid_data.flatten(order='F')[
-            grid_data.flatten(order='F') != -1
-            ]
-        reorder_index = reorder_index[:, None, None]  # [N] => [N, 1, 1]
-    else:
-        raise ValueError('Input cart_data must be a 1, 2, or 3 dimensional matrix.')
-
-    # compute the reverse ordering index (i.e., what is the index of each point
-    # in the reordering vector)
-    order_index = np.ones((reorder_index.size, 2), dtype=int)
-    order_index[:, 0] = np.squeeze(reorder_index)
-    order_index[:, 1] = np.arange(1, reorder_index.size + 1)
-    order_index = sort_rows(order_index, 0)
-    order_index = order_index[:, 1]
-    order_index = order_index[:, None]  # [N] => [N, 1]
-
-    # reset binary grid values
-    if kgrid.dim == 1:
-        grid_data[grid_data != 0] = 1
-    else:
-        grid_data[grid_data != -1] = 1
-        grid_data[grid_data == -1] = 0
-
-    # check if any Cartesian points have been mapped to the same grid point,
-    # thereby reducing the total number of points
-    num_discarded_points = cart_data.shape[1] - np.sum(grid_data)
-    if num_discarded_points != 0:
-        print(f'  cart2grid: {num_discarded_points} Cartesian points mapped to overlapping grid points')
-    return grid_data.astype(np.int), order_index, reorder_index
 
 
 def get_bli(func, dx=1, up_sampling_factor=20, plot=False):
