+++ kwave/utils/math.py
@@ -142,3 +142,78 @@ def primefactors(n):
         factors.append(n)
 
     return factors
+
+
+def next_pow2(n):
+    """
+    Calculate the next power of 2 that is greater than or equal to `n`.
+
+    This function takes a positive integer `n` and returns the smallest power of 2 that is greater
+    than or equal to `n`.
+
+    Args:
+        n: The number to find the next power of 2 for.
+
+    Returns:
+        The smallest power of 2 that is greater than or equal to `n`.
+    """
+
+    # decrement `n` (to handle cases when `n` itself is a power of 2)
+    n = n - 1
+
+    # set all bits after the last set bit
+    n |= n >> 1
+    n |= n >> 2
+    n |= n >> 4
+    n |= n >> 8
+    n |= n >> 16
+
+    # increment `n` and return
+    return n + 1
+
+
+def norm_var(im):
+    mu = np.mean(im)
+    s = np.sum((im - mu) ** 2) / mu
+    return s
+
+
+def gaussian(x, magnitude=None, mean=0, variance=1):
+    """
+    gaussian returns a Gaussian distribution f(x) with the specified
+    magnitude, mean, and variance. If these values are not specified, the
+    magnitude is normalised and values of variance = 1 and mean = 0 are
+    used. For example running
+
+        import matplotlib.pyplot as plt
+        x = np.arrange(-3,0.05,3)
+        plt.plot(x, gaussian(x))
+
+    will plot a normalised Gaussian distribution.
+
+    Note, the full width at half maximum of the resulting distribution
+    can be calculated by FWHM = 2 * sqrt(2 * log(2) * variance).
+
+
+    Args:
+        x:
+        magnitude:          Bell height. Defaults to normalized.
+        mean (float):       mean or expected value. Defaults to 0.
+        variance (float):   variance ~ bell width. Defaults to 1.
+
+    Returns:
+        gauss_distr: Gaussian distribution
+
+    """
+    if magnitude is None:
+        magnitude = (2 * math.pi * variance) ** -0.5
+
+    gauss_distr = magnitude * np.exp(-(x - mean) ** 2 / (2 * variance))
+
+    return gauss_distr
+    # return magnitude * norm.pdf(x, loc=mean, scale=variance)
+    """ # Former impl. form Farid
+        if magnitude is None:
+        magnitude = np.sqrt(2 * np.pi * variance)
+    return magnitude * np.exp(-(x - mean) ** 2 / (2 * variance))
+    """
