+++ tests/matlab_test_data_collectors/python_testers/cart2grid_test.py
@@ -5,7 +5,7 @@ from unittest.mock import Mock
 import numpy as np
 from scipy.io import loadmat
 
-from kwave.utils.interp import cart2grid
+from kwave.utils.conversion import cart2grid
 
 
 def test_cart2grid():
