+++ tests/matlab_test_data_collectors/python_testers/makeCircle_test.py
@@ -1,16 +1,10 @@
-from unittest.mock import Mock
-
-from kwave.utils.mapgen import make_circle
-
-from kwave.utils.interp import cart2grid
-
-from kwave.utils.conversion import scale_time
-
-from scipy.io import loadmat
-import numpy as np
 import os
 from pathlib import Path
-import pytest
+
+import numpy as np
+from scipy.io import loadmat
+
+from kwave.utils.mapgen import make_circle
 
 
 def test_makeCircle():
