+++ tests/test_na_optimising_performance.py
@@ -17,7 +17,7 @@ from kwave.kmedium import kWaveMedium
 from kwave.ksource import kSource
 from kwave.kspaceFirstOrder2D import kspaceFirstOrder2DC
 from kwave.ktransducer import kSensor
-from kwave.utils.interp import cart2grid
+from kwave.utils.conversion import cart2grid
 from kwave.utils.io import load_image
 from kwave.utils.mapgen import make_cart_circle
 from kwave.utils.matrix import resize
