+++ tests/test_sd_directivity_modelling_3D.py
@@ -19,8 +19,8 @@ from kwave.kmedium import kWaveMedium
 from kwave.ksensor import kSensor
 from kwave.ksource import kSource
 from kwave.kspaceFirstOrder3D import kspaceFirstOrder3DC
+from kwave.utils.conversion import cart2grid
 from kwave.utils.filters import filter_time_series
-from kwave.utils.interp import cart2grid
 from kwave.utils.mapgen import make_cart_circle
 from kwave.utils.matlab import matlab_find, unflatten_matlab_mask
 from tests.diff_utils import compare_against_ref
